+++ scheduler/pkg/server/pipeline_status_test.go
@@ -1,0 +1,114 @@
+/*
+Copyright (c) 2024 Seldon Technologies Ltd.
+
+Use of this software is governed by
+(1) the license included in the LICENSE file or
+(2) if the license included in the LICENSE file is the Business Source License 1.1,
+the Change License after the Change Date as each is defined in accordance with the LICENSE file.
+*/
+
+package server
+
+import (
+	"testing"
+	"time"
+
+	. "github.com/onsi/gomega"
+	log "github.com/sirupsen/logrus"
+
+	pb "github.com/seldonio/seldon-core/apis/go/v2/mlops/scheduler"
+
+	"github.com/seldonio/seldon-core/scheduler/v2/pkg/store/pipeline"
+)
+
+func TestPipelineStatusStream(t *testing.T) {
+	g := NewGomegaWithT(t)
+	type test struct {
+		name    string
+		loadReq *pb.LoadPipelineRequest
+		server  *SchedulerServer
+		timeout time.Duration
+		err     bool
+	}
+
+	tests := []test{
+		{
+			name: "pipeline ok",
+			loadReq: &pb.LoadPipelineRequest{
+				Pipeline: &pb.Pipeline{
+					Name:    "foo",
+					Version: 1,
+					Uid:     "x",
+					Steps: []*pb.PipelineStep{
+						{
+							Name: "a",
+						},
+						{
+							Name:   "b",
+							Inputs: []string{"a.outputs"},
+						},
+					},
+				},
+			},
+			server: &SchedulerServer{
+				pipelineHandler: pipeline.NewPipelineStore(log.New(), nil, nil),
+				logger:          log.New(),
+			},
+			timeout: 10 * time.Millisecond,
+		},
+		{
+			name: "timeout",
+			loadReq: &pb.LoadPipelineRequest{
+				Pipeline: &pb.Pipeline{
+					Name:    "foo",
+					Version: 1,
+					Uid:     "x",
+					Steps: []*pb.PipelineStep{
+						{
+							Name: "a",
+						},
+						{
+							Name:   "b",
+							Inputs: []string{"a.outputs"},
+						},
+					},
+				},
+			},
+			server: &SchedulerServer{
+				pipelineHandler: pipeline.NewPipelineStore(log.New(), nil, nil),
+				logger:          log.New(),
+			},
+			timeout: 1 * time.Millisecond,
+			err:     true,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			if test.loadReq != nil {
+				err := test.server.pipelineHandler.AddPipeline(test.loadReq.Pipeline)
+				g.Expect(err).To(BeNil())
+			}
+
+			stream := newStubPipelineStatusServer(1, 5*time.Millisecond)
+			err := test.server.sendCurrentPipelineStatuses(stream, false, test.timeout)
+			if test.err {
+				g.Expect(err).ToNot(BeNil())
+			} else {
+				g.Expect(err).To(BeNil())
+
+				var psr *pb.PipelineStatusResponse
+				select {
+				case next := <-stream.msgs:
+					psr = next
+				default:
+					t.Fail()
+				}
+
+				g.Expect(psr).ToNot(BeNil())
+				g.Expect(psr.Versions).To(HaveLen(1))
+				g.Expect(psr.Versions[0].State.Status).To(Equal(pb.PipelineVersionState_PipelineCreate))
+			}
+		})
+	}
+}
