+++ scheduler/pkg/server/server_test.go
@@ -654,246 +654,6 @@ func TestPipelineStatus(t *testing.T) {
 	}
 }
 
-func TestPipelineStatusStream(t *testing.T) {
-	g := NewGomegaWithT(t)
-	type test struct {
-		name    string
-		loadReq *pb.LoadPipelineRequest
-		server  *SchedulerServer
-		timeout time.Duration
-		err     bool
-	}
-
-	tests := []test{
-		{
-			name: "pipeline ok",
-			loadReq: &pb.LoadPipelineRequest{
-				Pipeline: &pb.Pipeline{
-					Name:    "foo",
-					Version: 1,
-					Uid:     "x",
-					Steps: []*pb.PipelineStep{
-						{
-							Name: "a",
-						},
-						{
-							Name:   "b",
-							Inputs: []string{"a.outputs"},
-						},
-					},
-				},
-			},
-			server: &SchedulerServer{
-				pipelineHandler: pipeline.NewPipelineStore(log.New(), nil, nil),
-				logger:          log.New(),
-			},
-			timeout: 10 * time.Millisecond,
-		},
-		{
-			name: "timeout",
-			loadReq: &pb.LoadPipelineRequest{
-				Pipeline: &pb.Pipeline{
-					Name:    "foo",
-					Version: 1,
-					Uid:     "x",
-					Steps: []*pb.PipelineStep{
-						{
-							Name: "a",
-						},
-						{
-							Name:   "b",
-							Inputs: []string{"a.outputs"},
-						},
-					},
-				},
-			},
-			server: &SchedulerServer{
-				pipelineHandler: pipeline.NewPipelineStore(log.New(), nil, nil),
-				logger:          log.New(),
-			},
-			timeout: 1 * time.Millisecond,
-			err:     true,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if test.loadReq != nil {
-				err := test.server.pipelineHandler.AddPipeline(test.loadReq.Pipeline)
-				g.Expect(err).To(BeNil())
-			}
-
-			stream := newStubPipelineStatusServer(1, 5*time.Millisecond)
-			err := test.server.sendCurrentPipelineStatuses(stream, false, test.timeout)
-			if test.err {
-				g.Expect(err).ToNot(BeNil())
-			} else {
-				g.Expect(err).To(BeNil())
-
-				var psr *pb.PipelineStatusResponse
-				select {
-				case next := <-stream.msgs:
-					psr = next
-				default:
-					t.Fail()
-				}
-
-				g.Expect(psr).ToNot(BeNil())
-				g.Expect(psr.Versions).To(HaveLen(1))
-				g.Expect(psr.Versions[0].State.Status).To(Equal(pb.PipelineVersionState_PipelineCreate))
-			}
-		})
-	}
-}
-
-func TestModelsStatusStream(t *testing.T) {
-	g := NewGomegaWithT(t)
-	type test struct {
-		name    string
-		loadReq *pb.LoadModelRequest
-		server  *SchedulerServer
-		timeout time.Duration
-		err     bool
-	}
-
-	tests := []test{
-		{
-			name: "model ok",
-			loadReq: &pb.LoadModelRequest{
-				Model: &pb.Model{
-					Meta: &pb.MetaData{Name: "foo"},
-				},
-			},
-			server: &SchedulerServer{
-				modelStore: store.NewMemoryStore(log.New(), store.NewLocalSchedulerStore(), nil),
-				logger:     log.New(),
-			},
-			timeout: 10 * time.Millisecond,
-		},
-		{
-			name: "timeout",
-			loadReq: &pb.LoadModelRequest{
-				Model: &pb.Model{
-					Meta: &pb.MetaData{Name: "foo"},
-				},
-			},
-			server: &SchedulerServer{
-				modelStore: store.NewMemoryStore(log.New(), store.NewLocalSchedulerStore(), nil),
-				logger:     log.New(),
-			},
-			timeout: 1 * time.Millisecond,
-			err:     true,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if test.loadReq != nil {
-				err := test.server.modelStore.UpdateModel(test.loadReq)
-				g.Expect(err).To(BeNil())
-			}
-
-			stream := newStubModelStatusServer(1, 5*time.Millisecond)
-			err := test.server.sendCurrentModelStatuses(stream, test.timeout)
-			if test.err {
-				g.Expect(err).ToNot(BeNil())
-			} else {
-				g.Expect(err).To(BeNil())
-
-				var msr *pb.ModelStatusResponse
-				select {
-				case next := <-stream.msgs:
-					msr = next
-				default:
-					t.Fail()
-				}
-
-				g.Expect(msr).ToNot(BeNil())
-				g.Expect(msr.Versions).To(HaveLen(1))
-				g.Expect(msr.Versions[0].State.State).To(Equal(pb.ModelStatus_ModelStateUnknown))
-			}
-		})
-	}
-}
-
-func TestModelsStatusEvents(t *testing.T) {
-	g := NewGomegaWithT(t)
-
-	createTestScheduler := func() (*SchedulerServer, *coordinator.EventHub) {
-		logger := log.New()
-		logger.SetLevel(log.WarnLevel)
-
-		eventHub, err := coordinator.NewEventHub(logger)
-		g.Expect(err).To(BeNil())
-
-		schedulerStore := store.NewMemoryStore(logger, store.NewLocalSchedulerStore(), eventHub)
-		experimentServer := experiment.NewExperimentServer(logger, eventHub, nil, nil)
-		pipelineServer := pipeline.NewPipelineStore(logger, eventHub, schedulerStore)
-
-		scheduler := scheduler2.NewSimpleScheduler(
-			logger,
-			schedulerStore,
-			scheduler2.DefaultSchedulerConfig(schedulerStore),
-		)
-		s := NewSchedulerServer(logger, schedulerStore, experimentServer, pipelineServer, scheduler, eventHub)
-
-		return s, eventHub
-	}
-	type test struct {
-		name    string
-		loadReq *pb.LoadModelRequest
-		timeout time.Duration
-	}
-
-	tests := []test{
-		{
-			name: "model ok",
-			loadReq: &pb.LoadModelRequest{
-				Model: &pb.Model{
-					Meta: &pb.MetaData{Name: "foo"},
-				},
-			},
-			timeout: 1 * time.Millisecond,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			s, hub := createTestScheduler()
-			if test.loadReq != nil {
-				err := s.modelStore.UpdateModel(test.loadReq)
-				g.Expect(err).To(BeNil())
-			}
-
-			stream := newStubModelStatusServer(1, test.timeout)
-			s.modelEventStream.streams[stream] = &ModelSubscription{
-				name:   "dummy",
-				stream: stream,
-				fin:    make(chan bool),
-			}
-			g.Expect(s.modelEventStream.streams[stream]).ToNot(BeNil())
-			hub.PublishModelEvent(modelEventHandlerName, coordinator.ModelEventMsg{
-				ModelName: "foo", ModelVersion: 1})
-
-			// to allow events to propagate
-			time.Sleep(500 * time.Millisecond)
-
-			var msr *pb.ModelStatusResponse
-			select {
-			case next := <-stream.msgs:
-				msr = next
-			default:
-				t.Fail()
-			}
-
-			g.Expect(msr).ToNot(BeNil())
-			g.Expect(msr.Versions).To(HaveLen(1))
-			g.Expect(msr.Versions[0].State.State).To(Equal(pb.ModelStatus_ModelStateUnknown))
-
-		})
-	}
-}
-
 type stubPipelineStatusServer struct {
 	msgs      chan *pb.PipelineStatusResponse
 	sleepTime time.Duration
